app/requirements.txt
flask
python-dotenv
bson
flask_restplus
Werkzeug
flask-cors
pytest